pyproject.toml
[project]
name = "ideas-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.13.0",
    "httpx>=0.28.1",
    "langchain-core>=1.2.12",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.8",
    "pgvector>=0.3.0",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=4.0.0",
    "langchain-openai>=0.3.0",
    "openai-agents>=0.0.7",
    "fastapi[standard]>=0.129.0",
    "markdown>=3.10.2",
]

[project.scripts]
brave-search = "brave_search.cli:main"
agent = "agent.cli:main"
research-loop = "research_loop.cli:main"
index-rag = "index_rag.cli:main"
insights = "insights.cli:main"
research-loop-openai = "research_loop_openai.cli:main"
pipeline = "pipeline.cli:main"
test-search = "test_search:main"
webapp = "webapp.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brave_search", "agent", "research_loop", "research_loop_openai", "index_rag", "insights", "pipeline", "webapp", "llm_factory.py", "test_search.py"]

[tool.pytest.ini_options]
markers = [
    "integration: e2e tests requiring external services (e.g. PostgreSQL)",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
